Track the availability at the device level
Rename LightMiddleware to LightDevice which is tracking the physical
light device. Multiple virtual lights operate on the same underlying
LightDevice instance.
A LightDevice holds onto a single sender, and will cache zone data
(while keeping track of whether zone data is updating, and the last
update time).

For this reason we want to keep track of the availability at the device
level instead of the entity level. Previously, if an entity became
unavailable and then skipped updates (because of the cache) we would
never flip the flag and the light would result uncontrollable.